added placeholder toolbars and layout
