+++ frontend/src/components/EditorLeftToolBar.jsx
@@ -1,0 +1,11 @@
+import '../css/EditorPageCSS/EditorLeftToolBar.css';
+
+const EditorLeftToolBar = () => {
+  return (
+    <div className='left-toolbar'>
+        <p>TODO TOOLBAR</p>
+    </div>
+  )
+}
+
+export default EditorLeftToolBar
