+++ frontend/src/css/App.css
@@ -1,3 +1,4 @@
 #root {
-  
+  /* allow sub components to use the whole screen width */
+  width: 100%;
 }
